+++ src/Sentry/Internal/FileSystemStub.cs
